requirements.txt
Flask==1.0.3
numpy==1.16.4
gunicorn==19.9.0
spacy==2.1.3
bert-extractive-summarizer